refactor(rust): extract schema analysis into db/schema.rs (slice 3/9) (#11)
Move M6 schema code (TypeCount/FieldStat/SchemaReport, bson_type_label,
infer_schema, analyze_schema_impl) out of lib.rs into a new db/ module
(db/schema.rs). Make connection_is_mock/require_real_client/mock_db
crate-visible for the db modules; lib.rs re-exports the schema items so
tests and the command wrapper are unchanged. cargo test green (60).
