+++ src-tauri/src/db/mod.rs
@@ -1,0 +1,3 @@
+//! Database-operation modules (split out of lib.rs).
+
+pub mod schema;
